+++ src/main/java/org/ureca/pinggubackend/domain/member/service/MemberService.java
@@ -1,0 +1,5 @@
+package org.ureca.pinggubackend.domain.member.service;
+
+public interface MemberService {
+    void deleteMember(long memberId);
+}
